requirements.txt
flask
python-dotenv
requests
openai
youtube-transcript-api
langchain
pytube